+++ SpotiFLAC/providers/gdstudio.py
@@ -14,7 +14,7 @@ from SpotiFLAC.core.models import DownloadResult, TrackMetadata
 from SpotiFLAC.core.musicbrainz import AsyncMBFetch, mb_result_to_tags
 from SpotiFLAC.core.tagger import EmbedOptions, embed_metadata_async
 
-from .base import BaseProvider
+from ..core.base import BaseProvider
 
 logger = logging.getLogger(__name__)
