+++ SpotiFLAC/providers/qobuz.py
@@ -43,7 +43,7 @@ from SpotiFLAC.core.signed_session_desktop import (
 )
 from SpotiFLAC.core.tagger import EmbedOptions, _print_mb_summary, embed_metadata_async
 
-from .base import BaseProvider
+from ..core.base import BaseProvider
 
 logger = logging.getLogger(__name__)
