--- telegram/docker-compose.yml
@@ -1,27 +1,0 @@
-services:
-  spotiflac:
-    image: ghcr.io/bartolomeorusso9/spotiflac-module-version:latest
-    container_name: spotiflac-bot
-    restart: unless-stopped
-    privileged: true
-    command: ["bot"]
-    shm_size: "1gb"
-    environment:
-      TELEGRAM_BOT_TOKEN: ${TELEGRAM_BOT_TOKEN}
-      # ALLOWED_USER_IDS: "123456789"
-      # QOBUZ_AUTH_TOKEN: ${QOBUZ_AUTH_TOKEN}
-    ports:
-      - "5900:5900"
-    volumes:
-      - ./.spotiflac_docker:/root/.spotiflac
-      - ./.cache_docker:/root/.cache/spotiflac
-      - ./downloads:/app/downloads   # or point the host side at your Jellyfin/Navidrome music folder
-  watchtower:
-    image: containrrr/watchtower:latest
-    container_name: watchtower_spotiflac
-    restart: unless-stopped
-    environment:
-      - DOCKER_API_VERSION=1.40  # Forza Watchtower a usare un'API aggiornata
-    volumes:
-      - /var/run/docker.sock:/var/run/docker.sock
-    command: --interval 300 --cleanup spotiflac
